requirements.txt
Django==5.2.1
djangorestframework_simplejwt==5.5.0
drf_spectacular==0.28.0
